Cleaned up files so far, workaround jquery/sass tooltipstyling
