+++ package.json
@@ -9,9 +9,10 @@
         "@vitejs/plugin-vue": "^4.0.0",
         "axios": "^1.1.2",
         "bootstrap": "^5.2.3",
+        "jquery": "^3.6.4",
         "laravel-vite-plugin": "^0.7.2",
         "sass": "^1.56.1",
-        "vite": "^4.0.0",
+        "vite": "^4.3.5",
         "vue": "^3.2.37"
     }
 }
