+++ resources/js/bootstrap.js
@@ -1,5 +1,5 @@
 import 'bootstrap'; 
-/* import 'jquery';  */
+import 'jquery'; 
 
 
 /**
