+++ resources/sass/_variables.scss
@@ -3,15 +3,18 @@ $body-bg: #f8fafc;
 
 // Typography
 $font-family-sans-serif: 'Nunito', sans-serif;
-$font-size-base: 0.9rem;
+$font-size-base: 1rem;
 $line-height-base: 1.6;
+
+// Cards
 $card-border-width: 5px;
 $card-border-radius: 20px;
+$card-height: null;
 
-$card-height:                       null;
-
-
+// Buttons
 $btn-border-radius-lg: 40px;
+
+// Tooltips
 $tooltip-color:#ffffff;
 $tooltip-bg: #905509;
 $tooltip-border-radius:10px;
