+++ resources/views/layouts/master.blade.php
@@ -10,18 +10,18 @@
 
     <title>{{ config('app.name', 'myOﬃceX') }}</title>
 
-
     <!-- Iconsevices -->
     <script src="https://kit.fontawesome.com/d4cbcb96c8.js" crossorigin="anonymous"></script>
     <!-- CSS-Animaitions -->
+
     <!-- Fonts -->
+
     <!-- Styles -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/css-only-tooltip@latest/dist/styles.min.css">
+    
     <!-- Scripts -->
-    {{-- <script src="{{ asset('node_modules/jquery/dist/jquery.min.js') }}"></script> --}}
 
     <!-- Directives -->
-    @vite(['resources/sass/app.scss', 'resources/css/app.css', 'resources/css/reset.css', 'resources/css/custom.css', 'resources/js/app.js']);
+    @vite(['resources/sass/app.scss', 'resources/css/app.css', 'resources/css/reset.css', 'resources/css/custom.css']);
 
 
 
@@ -42,10 +42,7 @@
     </div>
 
     <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-{{--  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
- <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNSxN9J" crossorigin="anonymous"></script>
- <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
- --}}
-
+<script src="https://code.jquery.com/jquery-3.6.4.min.js" integrity="sha256-oP6HI9z1XaZNBrJURtCoUT5SUnxFr8s3BzRl+cbzUq8=" crossorigin="anonymous"></script>
+@vite(['resources/js/app.js', 'resources/js/custom.js']);
 </body>
 </html>
