+++ resources/views/layouts/sectionIntro.blade.php
@@ -1,17 +1,17 @@
 <section class="view hm-gradient" id="intro">
-    <div class="site-bg-img d-flex align-items-center">
-      <div class="container">
-        <div class="row no-gutters">
-          <div class="col-md-10 col-lg-6 text-center text-md-left margins">
-            <div class="white-text">
-              <div class="wow fadeInLeft" data-wow-delay="0.3s">
-                <h1 class="h1-responsive font-weight-bold mt-sm-5">Büroraum, der auf deine Bedürfnisse eingeht.</h1>
-                <div class="h6"><b>Mit der myOfficeX-App können Mitarbeitende Feedback dazu geben, wie sie den Büroraum erleben. Auf Basis dieses Feedbacks generiert die App massgeschneiderte Empfehlungen zur Verbesserung des Arbeitsplatz-Komforts und der Energieeffizienz. </b></div>
-              </div><br>
-              <div class="wow fadeInLeft" data-wow-delay="0.3s"><a class="btn btn-white dark-grey-text font-weight-bold ml-0" href="https://www.1mai.ch/" target="_blank"><i class="fa fa-play mr-1"></i> Demo buchen</a><a class="btn btn-outline-white" href="white-paper.pdf"><b>White Paper</b></a></div>
-            </div>
+  <div class="site-bg-img d-flex align-items-center">
+    <div class="container">
+      <div class="row no-gutters">
+        <div class="col-md-10 col-lg-6 text-center text-md-left margins">
+          <div class="white-text">
+            <div class="wow fadeInLeft" data-wow-delay="0.3s">
+              <h1 class="h1-responsive font-weight-bold mt-sm-5">Büroraum, der auf deine Bedürfnisse eingeht.</h1>
+              <div class="h6"><b>Mit der myOfficeX-App können Mitarbeitende Feedback dazu geben, wie sie den Büroraum erleben. Auf Basis dieses Feedbacks generiert die App massgeschneiderte Empfehlungen zur Verbesserung des Arbeitsplatz-Komforts und der Energieeffizienz. </b></div>
+            </div><br>
+            <div class="wow fadeInLeft" data-wow-delay="0.3s"><a class="btn btn-white dark-grey-text font-weight-bold ml-0" href="https://www.1mai.ch/" target="_blank"><i class="fa fa-play mr-1"></i> Demo buchen</a><a class="btn btn-outline-white" href="white-paper.pdf"><b>White Paper</b></a></div>
           </div>
         </div>
       </div>
     </div>
-  </section>
+  </div>
+</section>
